Allow for kube clients to skip CRD creation (#32)
We want to be able to build kubernetes clients for users (tokens) with
limited privileges. Currently, the `Register` method tries to create
CRDs and fails in case the user is not authorized to perform that action.
I added a `skipCrdCreation` bool field to the kube factory to allow for
this action to be skipped. The bool's zero value makes this change
backwards compatible.